Fix bug with sleep times not showing for option 1
Comparison with mktime values didn't work because other than hour &
minute values, everything else is same between the 2. Wake_time was
initialized with values in cur_time to prevent errors from empty values.
Now, it only shows a time value if it is past the current time i.e. if
it is possible for you to sleep at that time. Creates new bug though #1